==> src/step2_prepare_taxation.Rmd <==
---
title: "Merge Taxation with Immunisation datasets"
output: html_document
author: Ethan- Quoc Do
---


```{r setup, include=FALSE}
#Load all neccessary library for this merge
library(dplyr)
library(readr)
library(tidyverse)
library(readxl)
options(scipen=999)

```

## Import Taxation

```{r}
#2016
tax_2016 <- read_xlsx("../raw_data/Taxation/2016.xlsx", sheet="Individuals Table 6B")

tax_2016 <- tax_2016 %>%
  rename("state"= "State/ Territory1") %>%
  rename("num_individuals"= "Number of individuals no.") %>%
  rename("total_tax"= "Taxable income or loss $") %>%
  rename("postcode"="Postcode") %>%
  select(state,postcode,num_individuals,total_tax)

tax_2016 <- tax_2016 %>%
  group_by(postcode) %>%
  summarise(total_tax = sum(total_tax),num_individuals = sum(num_individuals)) %>%
  mutate(mean_tax_000s = (ceiling(total_tax/num_individuals))/1000) %>%
  mutate(year = "2016")

tax_2016$total_tax <- as.numeric(tax_2016$total_tax)
tax_2016$num_individuals <- as.numeric(tax_2016$num_individuals)
tax_2016$postcode <- as.numeric(tax_2016$postcode)

tax_2016<-na.omit(tax_2016)

write.csv(tax_2016,"../raw_data/tax_2015.csv", row.names=FALSE)

```

```{r}
#2015
tax_2015 <- read_xlsx("../raw_data/Taxation/2015.xlsx", sheet="Individuals Table 6B")
#rename them first for easy to merge
tax_2015 <- tax_2015 %>%
  rename("state"= "State/ Territory1") %>%
  rename("num_individuals"= "Number of individuals\r\nno.") %>%
  rename("total_tax"= "Taxable income or loss3\r\n$") %>%
  rename("postcode"="Postcode") %>%
  select(state,postcode,num_individuals,total_tax)

tax_2015 <- tax_2015 %>%
  group_by(postcode) %>%
  summarise(total_tax = sum(total_tax),num_individuals = sum(num_individuals)) %>%
  mutate(mean_tax_000s = (ceiling(total_tax/num_individuals))/1000) %>%
  mutate(year = "2015")

tax_2015$total_tax <- as.numeric(tax_2015$total_tax)
tax_2015$num_individuals <- as.numeric(tax_2015$num_individuals)
tax_2015$postcode <- as.numeric(tax_2015$postcode)

tax_2015<-na.omit(tax_2015)

write.csv(tax_2015,"../raw_data/tax_2014.csv", row.names=FALSE)

```

```{r}
#2014
tax_2014 <- read_xlsx("../raw_data/Taxation/2014.xlsx", sheet="Table 6")

tax_2014 <- tax_2014 %>%
  rename("state"= "State/Territory1") %>%
  rename("num_individuals"= "Number of individuals") %>%
  rename("total_tax"= "Taxable income or loss3\r\n$") %>%
  rename("postcode"="Postcode") %>%
  select(state,postcode,num_individuals,total_tax)

tax_2014 <- tax_2014 %>%
  group_by(postcode) %>%
  summarise(total_tax = sum(total_tax),num_individuals = sum(num_individuals)) %>%
  mutate(mean_tax_000s = (ceiling(total_tax/num_individuals))/1000) %>%
  mutate(year = "2014")

tax_2014$total_tax <- as.numeric(tax_2014$total_tax)
tax_2014$num_individuals <- as.numeric(tax_2014$num_individuals)
tax_2014$postcode <- as.numeric(tax_2014$postcode)

tax_2014<-na.omit(tax_2014)

write.csv(tax_2014,"../raw_data/tax_2013.csv", row.names=FALSE)

```

```{r}
#2013
tax_2013 <- read_xlsx("../raw_data/Taxation/2013.xlsx", sheet="Postcode only")

tax_2013 <- tax_2013 %>%
  rename("state"= "State/Territory1") %>%
  rename("num_individuals"= "Number of individuals\r\nno.") %>%
  rename("total_tax"= "Taxable income or loss\r\n$") %>%
  rename("postcode"="Postcode") %>%
  select(state,postcode,num_individuals,total_tax)

tax_2013 <- tax_2013 %>%
  group_by(postcode) %>%
  summarise(total_tax = sum(total_tax),num_individuals = sum(num_individuals)) %>%
  mutate(mean_tax_000s = (ceiling(total_tax/num_individuals))/1000) %>%
  mutate(year = "2013")

tax_2013$total_tax <- as.numeric(tax_2013$total_tax)
tax_2013$num_individuals <- as.numeric(tax_2013$num_individuals)
tax_2013$postcode <- as.numeric(tax_2013$postcode)

tax_2013<-na.omit(tax_2013)

write.csv(tax_2013,"../raw_data/tax_2012.csv", row.names=FALSE)
  
```

```{r}
#2012
tax_2012 <- read_xlsx("../raw_data/Taxation/2012.xlsx", sheet="Individuals Tax Table 6")

tax_2012 <- tax_2012 %>%
  rename("state"= "State/Territory1") %>%
  rename("num_individuals"= "Number of individuals\r\nno.") %>%
  rename("total_tax"= "Taxable income or loss3\r\n$") %>%
  rename("postcode"="Postcode") %>%
  select(state,postcode,num_individuals,total_tax)

tax_2012 <- tax_2012 %>%
  group_by(postcode) %>%
  summarise(total_tax = sum(total_tax),num_individuals = sum(num_individuals)) %>%
  mutate(mean_tax_000s = (ceiling(total_tax/num_individuals))/1000) %>%
  mutate(year = "2012")

tax_2012$total_tax <- as.numeric(tax_2012$total_tax)
tax_2012$num_individuals <- as.numeric(tax_2012$num_individuals)
tax_2012$postcode <- as.numeric(tax_2012$postcode)

tax_2012<-na.omit(tax_2012)

write.csv(tax_2012,"../raw_data/tax_2011.csv", row.names=FALSE)
```
