==> src/step4_electoratescore.Rmd <==
---
title: "Step4_mergeelectorate"
author: "Alex Brooks"
date: "11 October 2018"
output: html_document
---
#Adding an election score
```{r setup, include=FALSE}
library(data.table)
library(readr)
library(dplyr)
library(tidyverse)
library(purrr)
```

## Bring in Step 3 cleaned data

 

```{r}
master_data <- read.csv("../cleaned_data/all_immunisation_seifa.csv")
pol_score <- read.csv("../raw_data/Transformed_elec.csv")

```

```{r}
#Get the master data for each year

master_data_2016 <- filter(master_data, year == 2016)
master_data_2015 <- filter(master_data, year == 2015)
master_data_2014 <- filter(master_data, year == 2014)
master_data_2013 <- filter(master_data, year == 2013)
master_data_2012 <- filter(master_data, year == 2012)
master_data_2011 <- filter(master_data, year == 2011)
 
```

```{r}
#Get the political score for each year

pol_score_2016 <- pol_score %>%
    select(postcode, X2016_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2016_score)*10))

pol_score_2015 <- pol_score %>%
    select(postcode, X2015_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2015_score)*10))

pol_score_2014 <- pol_score %>%
    select(postcode, X2014_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2014_score)*10))

pol_score_2013 <- pol_score %>%
    select(postcode, X2013_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2013_score)*10))

pol_score_2012 <- pol_score %>%
    select(postcode, X2012_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2012_score)*10))

pol_score_2011 <- pol_score %>%
    select(postcode, X2011_score) %>%
    group_by(postcode) %>%
    summarize(political_score = (mean(X2011_score)*10))
```

## Joining data

```{r}

combine.2016 <- merge(master_data_2016, pol_score_2016, by.x='postcode', by.y='postcode', all.x = TRUE)
combine.2015 <- merge(master_data_2015, pol_score_2015, by.x='postcode', by.y='postcode', all.x = TRUE)
combine.2014 <- merge(master_data_2014, pol_score_2014, by.x='postcode', by.y='postcode', all.x = TRUE)
combine.2013 <- merge(master_data_2013, pol_score_2013, by.x='postcode', by.y='postcode', all.x = TRUE)
combine.2012 <- merge(master_data_2012, pol_score_2012, by.x='postcode', by.y='postcode', all.x = TRUE)
combine.2011 <- merge(master_data_2011, pol_score_2011, by.x='postcode', by.y='postcode', all.x = TRUE)

```

```{r}
total.combined <- rbind(combine.2016, combine.2015, combine.2014, combine.2013, combine.2012, combine.2011)

write.csv(total.combined,"../cleaned_data/all_immunisation_seifa.csv", row.names=FALSE)
```
